Look up vlan IDs for bridge interfaces
Bridges usually sit on top of a vlan interface, so to figure out which
vlan they belong to we need to look up that interface, and check its
vlan configuration.

Due to limits of jinja variables this also requires rewriting the dns
name and vlan variables to use a dict.