Look up VLans on the physical node of a container
Containers only have a simple network interface with a link to an
interface of the physical node. The VLan configuration is either in that
link, or the interface the link sits on top.

As typically the container hooks into a bridge this only implements
looking up VLans under a bridge interface.

This commit also fixes a logic error skipping correct VLan assignments
if a DNS name was defined on the interface.