combustion: run in prepare as well, and provide default functions
Per default it'll resize root partition and add a separate var in the
initrd, and add root ssh keys after initrd.

If custom functions get added the defaults are not run - to make
things simpler all functions run by default are collected in two
functions, default_initrd and default_noinitrd. So to add a custom
snippet defining wipe_nvmes, and runn the predefined mirror_nvme_root
as well as the defaults a configuration like the follow would work:

  combustion:
    scriptlets:
      - combustion_wipenvmes.j2
    commands_initrd:
      - default_initrd
      - wipe_nvmes
    commands_noinitrd:
      - default_noinitrd
      - mirror_nvme_root